pyproject.toml
[project]
name = "Anocat-Backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Nurlan Urazkulov",email = "nurlan.urazkulov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn>=0.34.0,<0.35.0",
    "aiogram (>=3.18.0,<4.0.0)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "alembic>=1.15.1,<2.0.0",
    "sentry-sdk[fastapi] (>=2.22.0,<3.0.0)",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "tzdata>=2025.2",
]

[project.scripts]
app = "src.main:main"

[dependency-groups]
dev = []
